# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "Laravel API Swagger documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quizzes": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get authenticated user's quizzes",
                "operationId": "eb9ba15b35ba53ab91ba8b1a4e215e29",
                "responses": {
                    "200": {
                        "description": "List of quizzes"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Create a new quiz",
                "operationId": "2513a2f792e5c3a5bccd78eb9b4e3786",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "questions"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My Quiz"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "About Laravel"
                                    },
                                    "questions": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_text": {
                                                    "type": "string",
                                                    "example": "Laravel nima?"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "multiple_choice"
                                                },
                                                "options": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "option_text": {
                                                                "type": "string",
                                                                "example": "PHP freymvork"
                                                            },
                                                            "is_correct": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quiz created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/quizzes/{id}": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get one quiz by ID",
                "operationId": "d34cf9a999162f1e8ea1b8cce62d7cd5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz fetched successfully"
                    },
                    "404": {
                        "description": "Quiz not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/quizzes/{quizId}": {
            "post": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Update quiz",
                "operationId": "c2b003c1d9fe65f200fdd04690ab4b71",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "requestBody": {
                    "description": "Quiz update data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Quiz"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "questions": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_text": {
                                                    "type": "string",
                                                    "example": "Updated question text"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "multiple_choice"
                                                },
                                                "options": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "option_text": {
                                                                "type": "string",
                                                                "example": "Updated option text"
                                                            },
                                                            "is_correct": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Delete quiz",
                "operationId": "222a90065078af8dd7cfdd743b0c568c",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/all-quizzes": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get all quizzes with question count",
                "operationId": "6d54fc919d4655e9544310acb69da204",
                "responses": {
                    "200": {
                        "description": "All quizzes fetched successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/quiz/{quizId}/{questionId}": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get single question of a quiz",
                "operationId": "b05a28f303bcaa861b1cc44c3876ab0f",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question fetched"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/get-question-id/{quizId}": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get first question ID of a quiz",
                "operationId": "cd8005503c0d5524bb88078e6a8733cc",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "First question ID"
                    },
                    "404": {
                        "description": "Questions not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}